chore (ai): move ui-related files to src/ui
