+++ packages/ai/core/index.ts
@@ -19,10 +19,6 @@ export * from './util';
 // telemetry types:
 export type { TelemetrySettings } from './telemetry/telemetry-settings';
 
-// ui exports:
-export { getToolInvocations } from './ui/get-tool-invocations';
-export { getUIText } from './ui/get-ui-text';
-
 // util exports:
 export { cosineSimilarity } from './util/cosine-similarity';
 export { simulateReadableStream } from './util/simulate-readable-stream';
