+++ packages/ai/core/prompt/index.ts
@@ -1,5 +1,3 @@
-export { appendClientMessage } from './append-client-message';
-export { appendResponseMessages } from './append-response-messages';
 export type { CallSettings } from './call-settings';
 export type {
   FilePart,
@@ -8,10 +6,6 @@ export type {
   ToolCallPart,
   ToolResultPart,
 } from './content-part';
-export {
-  convertToCoreMessages,
-  convertToModelMessages,
-} from './convert-to-model-messages';
 export type { DataContent } from './data-content';
 export {
   assistantModelMessageSchema,
