+++ packages/ai/core/prompt/prompt.ts
@@ -1,4 +1,4 @@
-import { UIMessage } from '../types';
+import { UIMessage } from '../../src/ui/ui-messages';
 import { ModelMessage } from './message';
 
 /**
