--- packages/ai/core/types/ui-messages.ts
@@ -1,411 +1,0 @@
-import {
-  JSONValue,
-  LanguageModelV2FinishReason,
-  LanguageModelV2Source,
-} from '@ai-sdk/provider';
-import {
-  FetchFunction,
-  IdGenerator,
-  ToolCall,
-  ToolResult,
-} from '@ai-sdk/provider-utils';
-import { LanguageModelUsage } from './usage';
-
-/**
-Tool invocations are either tool calls or tool results. For each assistant tool call,
-there is one tool invocation. While the call is in progress, the invocation is a tool call.
-Once the call is complete, the invocation is a tool result.
-
-The step is used to track how to map an assistant UI message with many tool invocations
-back to a sequence of LLM assistant/tool result message pairs.
-It is optional for backwards compatibility.
- */
-export type ToolInvocation =
-  | ({ state: 'partial-call'; step?: number } & ToolCall<string, any>)
-  | ({ state: 'call'; step?: number } & ToolCall<string, any>)
-  | ({ state: 'result'; step?: number } & ToolResult<string, any, any>);
-
-/**
- * AI SDK UI Messages. They are used in the client and to communicate between the frontend and the API routes.
- */
-export interface UIMessage {
-  /**
-A unique identifier for the message.
-   */
-  id: string;
-
-  /**
-The timestamp of the message.
-   */
-  // TODO solve optionality similar id
-  createdAt?: Date;
-
-  /**
-The role of the message.
-   */
-  role: 'system' | 'user' | 'assistant';
-
-  /**
-Additional message-specific information added on the server via StreamData
-   */
-  // TODO replace with special part
-  annotations?: JSONValue[] | undefined;
-
-  /**
-The parts of the message. Use this for rendering the message in the UI.
-
-System messages should be avoided (set the system prompt on the server instead).
-They can have text parts.
-
-User messages can have text parts and file parts.
-
-Assistant messages can have text, reasoning, tool invocation, and file parts.
-   */
-  parts: Array<UIMessagePart>;
-}
-
-export type UIMessagePart =
-  | TextUIPart
-  | ReasoningUIPart
-  | ToolInvocationUIPart
-  | SourceUIPart
-  | FileUIPart
-  | StepStartUIPart;
-
-/**
- * A text part of a message.
- */
-export type TextUIPart = {
-  type: 'text';
-
-  /**
-   * The text content.
-   */
-  text: string;
-};
-
-/**
- * A reasoning part of a message.
- */
-export type ReasoningUIPart = {
-  type: 'reasoning';
-
-  /**
-   * The reasoning text.
-   */
-  text: string;
-
-  /**
-   * The provider metadata.
-   */
-  providerMetadata?: Record<string, any>;
-};
-
-/**
- * A tool invocation part of a message.
- */
-export type ToolInvocationUIPart = {
-  type: 'tool-invocation';
-
-  /**
-   * The tool invocation.
-   */
-  toolInvocation: ToolInvocation;
-};
-
-/**
- * A source part of a message.
- */
-export type SourceUIPart = {
-  type: 'source';
-
-  /**
-   * The source.
-   */
-  source: LanguageModelV2Source;
-};
-
-/**
- * A file part of a message.
- */
-export type FileUIPart = {
-  type: 'file';
-
-  /**
-   * IANA media type of the file.
-   *
-   * @see https://www.iana.org/assignments/media-types/media-types.xhtml
-   */
-  mediaType: string;
-
-  /**
-   * Optional filename of the file.
-   */
-  filename?: string;
-
-  /**
-   * The URL of the file.
-   * It can either be a URL to a hosted file or a [Data URL](https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/Data_URLs).
-   */
-  url: string;
-};
-
-/**
- * A step boundary part of a message.
- */
-export type StepStartUIPart = {
-  type: 'step-start';
-};
-
-export type CreateUIMessage = Omit<UIMessage, 'id'> & {
-  id?: UIMessage['id'];
-};
-
-export type ChatRequest = {
-  /**
-An optional object of headers to be passed to the API endpoint.
- */
-  headers?: Record<string, string> | Headers;
-
-  /**
-An optional object to be passed to the API endpoint.
-*/
-  body?: object;
-
-  /**
-The messages of the chat.
-   */
-  messages: UIMessage[];
-
-  /**
-Additional data to be sent to the server.
-   */
-  data?: JSONValue;
-};
-
-// Note: only used in useCompletion
-export type RequestOptions = {
-  /**
-An optional object of headers to be passed to the API endpoint.
- */
-  headers?: Record<string, string> | Headers;
-
-  /**
-An optional object to be passed to the API endpoint.
-   */
-  body?: object;
-};
-
-export type ChatRequestOptions = {
-  /**
-Additional headers that should be to be passed to the API endpoint.
- */
-  headers?: Record<string, string> | Headers;
-
-  /**
-Additional body JSON properties that should be sent to the API endpoint.
- */
-  body?: object;
-
-  /**
-Additional data to be sent to the API endpoint.
-   */
-  data?: JSONValue;
-
-  /**
-   * Allow submitting an empty message. Defaults to `false`.
-   */
-  allowEmptySubmit?: boolean;
-};
-
-export type UseChatOptions = {
-  /**
-   * The API endpoint that accepts a `{ messages: Message[] }` object and returns
-   * a stream of tokens of the AI chat response. Defaults to `/api/chat`.
-   */
-  api?: string;
-
-  /**
-   * A unique identifier for the chat. If not provided, a random one will be
-   * generated. When provided, the `useChat` hook with the same `id` will
-   * have shared states across components.
-   */
-  id?: string;
-
-  /**
-   * Initial messages of the chat. Useful to load an existing chat history.
-   */
-  initialMessages?: UIMessage[];
-
-  /**
-   * Initial input of the chat.
-   */
-  initialInput?: string;
-
-  /**
-Optional callback function that is invoked when a tool call is received.
-Intended for automatic client-side tool execution.
-
-You can optionally return a result for the tool call,
-either synchronously or asynchronously.
-   */
-  onToolCall?: ({
-    toolCall,
-  }: {
-    toolCall: ToolCall<string, unknown>;
-  }) => void | Promise<unknown> | unknown;
-
-  /**
-   * Callback function to be called when the API response is received.
-   */
-  onResponse?: (response: Response) => void | Promise<void>;
-
-  /**
-   * Optional callback function that is called when the assistant message is finished streaming.
-   *
-   * @param message The message that was streamed.
-   * @param options.usage The token usage of the message.
-   * @param options.finishReason The finish reason of the message.
-   */
-  onFinish?: (
-    message: UIMessage,
-    options: {
-      usage: LanguageModelUsage;
-      finishReason: LanguageModelV2FinishReason;
-    },
-  ) => void;
-
-  /**
-   * Callback function to be called when an error is encountered.
-   */
-  onError?: (error: Error) => void;
-
-  /**
-   * A way to provide a function that is going to be used for ids for messages and the chat.
-   * If not provided the default AI SDK `generateId` is used.
-   */
-  generateId?: IdGenerator;
-
-  /**
-   * The credentials mode to be used for the fetch request.
-   * Possible values are: 'omit', 'same-origin', 'include'.
-   * Defaults to 'same-origin'.
-   */
-  credentials?: RequestCredentials;
-
-  /**
-   * HTTP headers to be sent with the API request.
-   */
-  headers?: Record<string, string> | Headers;
-
-  /**
-   * Extra body object to be sent with the API request.
-   * @example
-   * Send a `sessionId` to the API along with the messages.
-   * ```js
-   * useChat({
-   *   body: {
-   *     sessionId: '123',
-   *   }
-   * })
-   * ```
-   */
-  body?: object;
-
-  /**
-Streaming protocol that is used. Defaults to `data`.
-   */
-  streamProtocol?: 'data' | 'text';
-
-  /**
-Custom fetch implementation. You can use it as a middleware to intercept requests,
-or to provide a custom fetch implementation for e.g. testing.
-    */
-  fetch?: FetchFunction;
-
-  /**
-Maximum number of sequential LLM calls (steps), e.g. when you use tool calls.
-Must be at least 1.
-
-A maximum number is required to prevent infinite loops in the case of misconfigured tools.
-
-By default, it's set to 1, which means that only a single LLM call is made.
- */
-  maxSteps?: number;
-};
-
-export type UseCompletionOptions = {
-  /**
-   * The API endpoint that accepts a `{ prompt: string }` object and returns
-   * a stream of tokens of the AI completion response. Defaults to `/api/completion`.
-   */
-  api?: string;
-  /**
-   * An unique identifier for the chat. If not provided, a random one will be
-   * generated. When provided, the `useChat` hook with the same `id` will
-   * have shared states across components.
-   */
-  id?: string;
-
-  /**
-   * Initial prompt input of the completion.
-   */
-  initialInput?: string;
-
-  /**
-   * Initial completion result. Useful to load an existing history.
-   */
-  initialCompletion?: string;
-
-  /**
-   * Callback function to be called when the API response is received.
-   */
-  onResponse?: (response: Response) => void | Promise<void>;
-
-  /**
-   * Callback function to be called when the completion is finished streaming.
-   */
-  onFinish?: (prompt: string, completion: string) => void;
-
-  /**
-   * Callback function to be called when an error is encountered.
-   */
-  onError?: (error: Error) => void;
-
-  /**
-   * The credentials mode to be used for the fetch request.
-   * Possible values are: 'omit', 'same-origin', 'include'.
-   * Defaults to 'same-origin'.
-   */
-  credentials?: RequestCredentials;
-
-  /**
-   * HTTP headers to be sent with the API request.
-   */
-  headers?: Record<string, string> | Headers;
-
-  /**
-   * Extra body object to be sent with the API request.
-   * @example
-   * Send a `sessionId` to the API along with the prompt.
-   * ```js
-   * useChat({
-   *   body: {
-   *     sessionId: '123',
-   *   }
-   * })
-   * ```
-   */
-  body?: object;
-
-  /**
-Streaming protocol that is used. Defaults to `data`.
-   */
-  streamProtocol?: 'data' | 'text';
-
-  /**
-Custom fetch implementation. You can use it as a middleware to intercept requests,
-or to provide a custom fetch implementation for e.g. testing.
-    */
-  fetch?: FetchFunction;
-};
