+++ packages/ai/src/index.ts
@@ -1,3 +1,4 @@
 // directory exports
 export * from './data-stream';
 export * from './text-stream';
+export * from './ui';
