+++ packages/ai/src/ui/convert-file-list-to-file-ui-parts.ts
@@ -1,4 +1,4 @@
-import { FileUIPart } from '../types';
+import { FileUIPart } from './ui-messages';
 
 export async function convertFileListToFileUIParts(
   files: FileList | undefined,
