+++ packages/ai/src/ui/extract-max-tool-invocation-step.ts
@@ -1,4 +1,4 @@
-import { ToolInvocation } from '../types';
+import { ToolInvocation } from './ui-messages';
 
 export function extractMaxToolInvocationStep(
   toolInvocations: ToolInvocation[] | undefined,
