+++ packages/ai/src/ui/get-ui-text.ts
@@ -1,4 +1,4 @@
-import { UIMessagePart } from '../types/ui-messages';
+import { UIMessagePart } from './ui-messages';
 
 export function getUIText(parts: UIMessagePart[]): string {
   return parts.map(part => (part.type === 'text' ? part.text : '')).join('');
