+++ packages/ai/src/ui/process-chat-text-response.ts
@@ -1,7 +1,8 @@
 import { JSONValue } from '@ai-sdk/provider';
 import { generateId as generateIdFunction } from '@ai-sdk/provider-utils';
 import { processTextStream } from './process-text-stream';
-import { TextUIPart, UIMessage, UseChatOptions } from '../types';
+import { TextUIPart, UIMessage } from './ui-messages';
+import { UseChatOptions } from './use-chat';
 
 export async function processChatTextResponse({
   stream,
