+++ packages/ai/src/ui/process-text-stream.ts
